+++ lib/kitchen/directions/bake_chapter_section_exercises/main.rb
@@ -5,11 +5,13 @@ module Kitchen
     module BakeChapterSectionExercises
       def self.v1(chapter:, options: {
         trash_title: false,
-        create_title: true
+        create_title: true,
+        numbering_options: { mode: :chapter_page, separator: '.' }
       })
         options.reverse_merge!(
           trash_title: false,
-          create_title: true
+          create_title: true,
+          numbering_options: { mode: :chapter_page, separator: '.' }
         )
         V1.new.bake(chapter: chapter, options: options)
       end